+++ sososhopping_customer/.idea/misc.xml
@@ -34,6 +34,7 @@
         <entry key="..\:/Hyunsik/program/AndroidProgramming/sososhopping_app_customer/sososhopping_customer/app/src/main/res/layout/account_log_in_dialog.xml" value="0.30570652173913043" />
         <entry key="..\:/Hyunsik/program/AndroidProgramming/sososhopping_app_customer/sososhopping_customer/app/src/main/res/layout/activity_main.xml" value="0.29438405797101447" />
         <entry key="..\:/Hyunsik/program/AndroidProgramming/sososhopping_app_customer/sososhopping_customer/app/src/main/res/layout/chat_list.xml" value="0.36614583333333334" />
+        <entry key="..\:/Hyunsik/program/AndroidProgramming/sososhopping_app_customer/sososhopping_customer/app/src/main/res/layout/conversation.xml" value="0.36614583333333334" />
         <entry key="..\:/Hyunsik/program/AndroidProgramming/sososhopping_app_customer/sososhopping_customer/app/src/main/res/layout/error_ask_log_in_dialog.xml" value="0.30570652173913043" />
         <entry key="..\:/Hyunsik/program/AndroidProgramming/sososhopping_app_customer/sososhopping_customer/app/src/main/res/layout/home.xml" value="0.29438405797101447" />
         <entry key="..\:/Hyunsik/program/AndroidProgramming/sososhopping_app_customer/sososhopping_customer/app/src/main/res/layout/item_cart_item.xml" value="0.3283514492753623" />
