+++ sososhopping_customer/app/build.gradle
@@ -81,10 +81,10 @@ dependencies {
     // Retrofit2 dependencies
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-    implementation 'com.google.code.gson:gson:2.8.7'
 
     //NaverMap
     implementation 'com.google.android.gms:play-services-location:16.0.0'
+    implementation 'com.google.code.gson:gson:2.8.7'
     implementation 'com.naver.maps:map-sdk:3.13.0'
 
     //dateTimePicker
