--- sososhopping_customer/app/src/main/java/com/sososhopping/customer/common/chat/ChatroomInfor.java
@@ -1,22 +1,0 @@
-package com.sososhopping.customer.common.chat;
-
-public class ChatroomInfor {
-    public String customerName = "";
-    public String storeName;
-    public String chatroomId = "";
-    public String lastMessage;
-    public Object lastMessageTimestamp;
-    public Object leaveChatroomTimestamp;
-
-    public ChatroomInfor() {
-    }
-
-    public ChatroomInfor(String customerName, String storeName, String chatroomId, String lastMessage, Object lastMessageTimestamp, Object leaveChatroomTimestamp) {
-        this.customerName = customerName;
-        this.storeName = storeName;
-        this.chatroomId = chatroomId;
-        this.lastMessage = lastMessage;
-        this.lastMessageTimestamp = lastMessageTimestamp;
-        this.leaveChatroomTimestamp = leaveChatroomTimestamp;
-    }
-}
