+++ sososhopping_customer/app/src/main/java/com/sososhopping/customer/interest/view/InterestShopListFragment.java
@@ -125,10 +125,7 @@ public class InterestShopListFragment extends Fragment {
     public void onResume() {
         //상단바
         ((HomeActivity)getActivity()).showTopAppBar();
-        ((HomeActivity)getActivity()).getBinding().topAppBar.setTitle("관심가게");
-        ((HomeActivity)getActivity()).getBinding().topAppBar.setTitleCentered(true);
-        ((HomeActivity)getActivity()).getBinding().topAppBar.setOnClickListener(null);
-        ((HomeActivity)getActivity()).setTopAppBarHome(false);
+        ((HomeActivity)getActivity()).setTopAppBarHome("관심가게");
 
         //하단바
         ((HomeActivity)getActivity()).showBottomNavigation();
