+++ sososhopping_customer/app/src/main/res/layout/search_shop_map.xml
@@ -28,7 +28,7 @@
             android:padding="10dp"
             app:cardBackgroundColor="#D9FFFFFF"
             app:cardCornerRadius="15dp"
-            app:cardElevation="5dp"
+            app:cardElevation="10dp"
             app:rippleColor="@color/white">
 
             <TextView
@@ -47,7 +47,7 @@
             android:clickable="true"
             app:cardBackgroundColor="#D9FFFFFF"
             app:cardCornerRadius="27.5dp"
-            app:cardElevation="5dp"
+            app:cardElevation="10dp"
             app:rippleColor="@color/white">
 
             <ImageView
