+++ sososhopping_customer/app/src/main/res/navigation/nav_graph.xml
@@ -45,6 +45,7 @@
         <action
             android:id="@+id/action_shopListFragment_to_shopMapFragment"
             app:destination="@id/shopMapFragment"
+            app:launchSingleTop="true"
             app:popUpTo="@id/shopListFragment"
             app:popUpToInclusive="false" />
         <action
@@ -84,7 +85,7 @@
         <action
             android:id="@+id/action_shopMapFragment_to_shopListFragment"
             app:destination="@id/shopListFragment"
-            app:popUpTo="@id/shopMapFragment"
+            app:popUpTo="@id/shopListFragment"
             app:popUpToInclusive="true" />
     </fragment>
 
@@ -330,16 +331,32 @@
         tools:layout="@layout/mysoso_setting"/>
     <fragment
         android:id="@+id/chatFragment"
-        android:name="com.sososhopping.customer.common.chat.view.ChatFragment"
+        android:name="com.sososhopping.customer.chat.view.ChatFragment"
         android:label="ChatFragment"
         tools:layout="@layout/chat_list">
         <action
             android:id="@+id/action_chatFragment_to_conversationFragment"
             app:destination="@id/conversationFragment" />
     </fragment>
+
     <fragment
         android:id="@+id/conversationFragment"
-        android:name="com.sososhopping.customer.common.chat.view.ConversationFragment"
-        android:label="ConversationFragment" />
+        android:name="com.sososhopping.customer.chat.view.ConversationFragment"
+        android:label="ConversationFragment"
+        tools:layout="@layout/conversation">
+        <argument
+            android:name="storeName"
+            app:argType="string"
+            app:nullable="true" />
+        <argument
+            android:name="storeId"
+            app:argType="long"
+            android:defaultValue="-1L" />
+        <argument
+            android:name="ownerId"
+            app:argType="long"
+            android:defaultValue="-1L" />
+    </fragment>
+    <action android:id="@+id/action_global_conversationFragment" app:destination="@id/conversationFragment" />
 
 </navigation>
